Move the UI operations and timestring generation to a new class in files.ts
